01_zmq/requirements.txt
pyzmq==25.1.2 